refactor EventFilter
