+++ ampd/src/event_processor.rs
@@ -15,7 +15,7 @@ use valuable::Valuable;
 
 use crate::asyncutil::future::{with_retry, RetryPolicy};
 use crate::asyncutil::task::TaskError;
-use crate::grpc::reqs::EventFilters;
+use crate::event_sub::event_filter::EventFilters;
 use crate::monitoring::metrics;
 use crate::monitoring::metrics::{Msg, Stage};
 use crate::{broadcast, cosmos, event_sub, monitoring};
@@ -276,7 +276,7 @@ mod tests {
     use crate::broadcast::test_utils::create_base_account;
     use crate::broadcast::DecCoin;
     use crate::event_processor::{consume_events, Config, Error, EventHandler};
-    use crate::grpc::reqs::EventFilters;
+    use crate::event_sub::event_filter::EventFilters;
     use crate::types::{random_cosmos_public_key, TMAddress};
     use crate::{broadcast, cosmos, event_sub, monitoring, PREFIX};
